Fix dataview tables having extra spacing
The selector looked for divs without classes to signify reading mode.
This caused issues with dataview tables which have their tables within
a div even in live preview.

To fix this, an extra class check has been added to make sure that the
margin only appears when the page itself is already in reading mode.