Toggle to enable H2 heading underlines
I love the theme but I find my notes much easier to read when the H2 headings have underlines.  If you're cool with it, I'd like to add the toggleable option to have underlines under H2 headings with this commit.  
I've left the option off by default so as to not affect users who do not wish for the change/option.